+++ src/py/altium_cruncher/_cli.py
@@ -71,6 +71,9 @@ from altium_cruncher.altium_cruncher_cmd_mco import (
 from altium_cruncher.altium_cruncher_cmd_notes import (
     register_parser as register_notes_parser,
 )
+from altium_cruncher.altium_cruncher_cmd_outjob import (
+    register_parser as register_outjob_parser,
+)
 from altium_cruncher.altium_cruncher_cmd_pcb_svg import (
     register_parser as register_pcb_svg_parser,
 )
@@ -259,6 +262,7 @@ def main() -> None:
     register_megamaid_parser(command_subparsers)
     register_merge_parser(command_subparsers)
     register_notes_parser(command_subparsers)
+    register_outjob_parser(command_subparsers)
     register_pcb_layer_step_parser(command_subparsers)
     register_pcb_svg_parser(command_subparsers)
     register_pnp_parser(command_subparsers)